Implemented Word Duel's core logic in the ActiveGameScreen, SMS based communication with an initial handshake process to prevent spoofing and as a simple means of verification.
Implemented a SimpleHandshakeHandler class to act as an interface for the handshake process, as well as the construction of valid messages the application will recognize and respond to.

Implemented the BroadcastReceiver is its own class.

Added a JoinMatchScreen. Here, players wait for an INV, which beings the handshake process, before being sent to the ActiveGameScreen.

Implemented a GameState class that collects details about an ongoing match. This object is then appended to a game_history file as a record of all games a player had engaged in.

Implemented a HighScore screen that read the data in game_history and display it in a simple ListView.

Updated the GameOverScreen. It displays the final score of each player and states who won the match.